Correct README credits, finish section-3
A container was added in section 3 to center the quote more nicely.
Before, the text would fill the entire width of the page, and only then
would it go to the next line. This time, because of the container, text
will only fill up to the container's width (set with flex-basis: 800px)
before it has to go into a new line, making this section look a lot
nicer to read.